move some code to a new arrow_traits.h and arrow_util.h files
