+++ cpp-client/deephaven/client/include/private/deephaven/client/arrowutil/arrow_value_converter.h
@@ -11,18 +11,26 @@
 namespace deephaven::client::arrowutil {
 class ArrowValueConverter {
 public:
-  // "convert" function for anything except string_view is the identity function.
+  /**
+   * The default "convert" function is the identity function.
+   */
   template<typename SRC, typename DEST>
   static void convert(SRC src, DEST *dest) {
     *dest = src;
   }
 
-  // "convert" function for string_view is std::string
+  /**
+   * The "convert" function for string_view is std::string
+   */
   static void convert(arrow::util::string_view sv, std::string *dest) {
     dest->clear();
     dest->append(sv.data(), sv.size());
   }
 
+  /**
+   * The "convert" function for the nanos-since-epoch representation of timestamps is
+   * deephaven::client::DateTime.
+   */
   static void convert(int64_t src, deephaven::client::DateTime *dest) {
     *dest = deephaven::client::DateTime::fromNanos(src);
   }
