+++ cpp-client/deephaven/client/include/private/deephaven/client/arrowutil/arrow_visitors.h
@@ -123,29 +123,4 @@ public:
 private:
   Inner inner_;
 };
-
-/**
- * Returns true iff type T is one of the numeric types.
- */
-template<typename T>
-constexpr bool isNumericType() {
-  static_assert(
-      std::is_same_v<T, int8_t> ||
-      std::is_same_v<T, int16_t> ||
-      std::is_same_v<T, int32_t> ||
-      std::is_same_v<T, int64_t> ||
-      std::is_same_v<T, float> ||
-      std::is_same_v<T, double> ||
-      std::is_same_v<T, bool> ||
-      std::is_same_v<T, std::string> ||
-      std::is_same_v<T, deephaven::client::DateTime>,
-          "T is not one of the supported element types for Deephaven columns");
-
-  return std::is_same_v<T, int8_t> ||
-      std::is_same_v<T, int16_t> ||
-      std::is_same_v<T, int32_t> ||
-      std::is_same_v<T, int64_t> ||
-      std::is_same_v<T, float> ||
-      std::is_same_v<T, double>;
-}
 }  // namespace deephaven::client::arrowutil
